requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
yt-dlp>=2024.1.0
python-multipart>=0.0.6
itsdangerous>=2.1.2
bcrypt>=4.0.0
python-dotenv>=1.0.0
